+++ HISTORY.rst
@@ -2,6 +2,21 @@
 History
 =======
 
+0.3.0 (2020-04-10)
+------------------
+
+* Release with merge for standard BAM maf and Input MAF. Converted multiple-patient to multiple-sample
+
+0.2.1 (2020-04-09)
+------------------
+
+* Release bug fixes, where simplex numbers are listed as duplex and vice versa, during running `all` command.
+
+0.2.0 (2020-04-08)
+------------------
+
+* Release with multiple-patient command.
+
 0.1.0 (2020-01-30)
 ------------------